pyproject.toml
[tool.poetry]
name = "llama2-blog-generation"
version = "0.1.0"
description = ""
authors = ["Adrian Moses <adrian@adrianmoses.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
ollama = "^0.1.6"
sentence-transformers = "^2.3.1"
ctransformers = "^0.2.27"
langchain = "^0.1.5"
langchain-community = "^0.0.17"
streamlit = "^1.31.0"
transformers = "^4.37.2"
huggingface-hub = "^0.20.3"
watchdog = "^3.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
